#9744: updated cpp references to refer to ttnn pad header
