+++ tt_eager/tt_dnn/op_library/composite/composite_ops.cpp
@@ -18,7 +18,6 @@
 #include "tt_dnn/op_library/unpad/unpad_op.hpp"
 #include "tt_eager/tensor/tensor_impl.hpp"
 #include "tt_eager/tensor/tensor_utils.hpp"
-#include "tt_eager/tt_dnn/op_library/pad/pad_op.hpp"
 #include "tt_eager/tt_dnn/op_library/unpad/unpad_op.hpp"
 #include "tt_numpy/functions.hpp"
 
@@ -1394,8 +1393,17 @@ Tensor hypot(const Tensor& input_a, const Tensor& input_b, const MemoryConfig& o
 
 Tensor _scatter(const Tensor& input_a, const Tensor& input_b, const MemoryConfig& output_mem_config) {
     const Shape start_index = {0, 0, 0, 0};
-    Tensor index = pad(ones_like(input_a, output_mem_config), input_b.get_legacy_shape(), start_index, 0);
-    Tensor temp_a = pad(input_a, input_b.get_legacy_shape(), start_index, 0);
+    ttnn::Tensor input_tensor_4D = ttnn::unsqueeze_to_4D(input_a);        
+    auto input_shape_with_tile_padding = input_tensor_4D.get_shape().with_tile_padding();
+    auto output_padded_shape = input_b.legacy_shape();
+    std::vector<std::pair<uint32_t, uint32_t>> padding(4); 
+    for(size_t i = 0; i < padding.size(); i++) {
+        padding[i] = {0, output_padded_shape[i] - input_shape_with_tile_padding[i]};
+    }
+
+
+    Tensor index = ttnn::pad(ones_like(input_a, output_mem_config), padding, 0, std::nullopt);
+    Tensor temp_a = ttnn::pad(input_a, padding, 0, std::nullopt);
     return where(index, temp_a, input_b, output_mem_config);
 }
 Tensor scatter(const Tensor& input_a, const Tensor& input_b, const MemoryConfig& output_mem_config) {
